Add RAPT caching & scopes for narrowing challenges in reauth flows.
- Cache RAPT to prevent frequent reauth challenges.

- This plugin is mainly used by gsutil, or with refresh tokens obtained
through gsutil (see instructions at [1]), and thus we expect refresh
tokens to usually have the cloud-platform API scope. We set this in
the default RAPT scopes, but we define the defaults at module level so
they may be easily overridden for use in other contexts as necessary.

Context from issue tracker:
https://issuetracker.google.com/issues/35099526#comment52

[1] https://cloud.google.com/storage/docs/boto-plugin